apps/postgres_to_es/migrator/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
elastic-transport==8.17.1
elasticsearch==8.6.2
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
